migrating js code into seperate files
