+++ src/components/questions/questionsPrb.js
@@ -1,0 +1,12 @@
+export default {
+    props:{
+        problemContainer:{
+           type:Object
+        }
+    },
+
+    method:{
+      
+    }
+
+}
